+++ src/Timer.js
@@ -1,191 +1,38 @@
-// import { Component } from 'react';
-// import { CircularProgressbarWithChildren } from 'react-circular-progressbar';
-// import 'react-circular-progressbar/dist/styles.css';
-// import './Timer.css'
-// import {secToHMS} from './tools'
-
-// class Timer extends Component {
-//   constructor(props) {
-//     super(props);
-//     this.secondsPerPayment = props.secondsPerPayment;
-//     this.pricePerInterval = props.pricePerinterval;
-
-//     this.state = {
-//       seconds: 0,
-//       isTimerRunning: false,
-//     };
-
-//   }
-
-//   getAccumulativePrice() {
-//     const amountOfIntervals = Math.floor(this.state.seconds / this.secondsPerPayment);
-//     const accumulativePrice = amountOfIntervals * this.pricePerInterval
-//     return accumulativePrice
-//   }
-
-//   getTimeFormated = (timeSeconds, useColons) => {
-
-//     const { h, m, s } = secToHMS(timeSeconds)
-
-//     let formatedString = "";
-//     // convert to string and 0 pad
-//     // different versions of formating
-//     if (useColons === true) {
-//       const [stringHours, stringMinutes, stringSeconds] = [h, m, s].map(value => value.toString().padStart(2, '0'))
-//       formatedString = `${stringHours}:${stringMinutes}:${stringSeconds}`
-//     } else {
-//       if (h !== 0) {
-//         formatedString += `${h}h`
-//       } if (m !== 0) {
-//         formatedString += `${m}m`
-//       } if (s !== 0) {
-//         formatedString += `${s}s`
-//       }
-//     }
-//     return formatedString
-//   }
-
-//   getPercentage() {
-//     const percentage = (this.state.seconds % this.secondsPerPayment) / this.secondsPerPayment
-//     return percentage
-//   }
-
-//   onTimerUpdate = () => {
-//     // update timer by 1
-//     this.setState({ seconds: this.state.seconds + 1 });
-//   }
-//   onTimerPlay = () => {
-//     // set the timers intervals
-//     this.f = setInterval(this.onTimerUpdate, 1000);
-//     this.setState({ isTimerRunning: true })
-//   }
-//   onTimerPause = () => {
-//     // clear the timer from ticking
-//     clearInterval(this.f);
-//     this.setState({ isTimerRunning: false })
-//   }
-//   onTimerReset = () => {
-//     this.onTimerPause()
-//     // resets the timer 
-//     this.setState({ seconds: 0 })
-//   }
-
-//   render() {
-//     return (
-//       <div className='timer'>
-//         <CircularProgressbarWithChildren value={this.getPercentage()} maxValue={1}
-//           styles={{
-//             path: {
-//               stroke: 'rgba(62, 152, 199, 255)',
-//               strokeLinecap: 'butt',
-//               transition: 'stroke-dashoffset 0.1s ease 0s',
-//             },
-//             trail: {
-//               stroke: '#333',
-//               strokeLinecap: 'butt',
-//             },
-//           }}>
-//           <p>{this.getTimeFormated(this.state.seconds, true)}</p>
-//           <p>{`£${this.getAccumulativePrice()} @ £${this.pricePerInterval}/${this.getTimeFormated(this.secondsPerPayment, false)}`}</p>
-//           <button id='btnPause' onClick={this.onTimerPause} disabled={!this.state.isTimerRunning}>pause</button>
-//           <button id='btnPlay' onClick={this.onTimerPlay} disabled={this.state.isTimerRunning}>play</button>
-//           <button id='btnReset' onClick={this.onTimerReset}>reset</button>
-
-//         </CircularProgressbarWithChildren>
-//       </div>
-//     )
-//   }
-// }
-
-// export default Timer
-
-
-
-
-import { Component } from 'react';
-import { CircularProgressbarWithChildren } from 'react-circular-progressbar';
-import 'react-circular-progressbar/dist/styles.css';
+import { useState } from 'react';
 import './Timer.css'
-import {secToHMS, getTimeFormated} from './tools'
+import TimerForm from './TimerForm';
+import InnerTimer from './InnerTimer';
 
-class Timer extends Component {
-  constructor(props) {
-    super(props);
-    this.secondsPerPayment = props.secondsPerPayment;
-    this.pricePerInterval = props.pricePerinterval;
+function Timer() {
 
-    this.state = {
-      seconds: 0,
-      isTimerRunning: false,
-    };
+  const [seconds, setSeconds] = useState(0)
+  const [isTimerRunning, setIsTimerRunning] = useState(false)
+  const [confirmed, setConfirmed] = useState(false)
+  const [pricePerinterval, setPricePerinterval] = useState(10.0)
+  const [secondsPerPayment, setSecondsPerPayment] = useState(3600)
 
-  }
+  // this.state = {
+  //   seconds: 0,
+  //   isTimerRunning: false,
+  //   confirmed: false,
+  //   pricePerinterval: 10.0,
+  //   secondsPerPayment: 3600,
+  // };
 
-  getAccumulativePrice() {
-    const amountOfIntervals = Math.floor(this.state.seconds / this.secondsPerPayment);
-    const accumulativePrice = (amountOfIntervals * this.pricePerInterval) + this.pricePerInterval
-    return accumulativePrice
-  }
+  return (
 
-  getPercentage() {
-    const percentage = (this.state.seconds % this.secondsPerPayment) / this.secondsPerPayment
-    return percentage
-  }
+    confirmed
+      ?
+      <InnerTimer
+        pricePerinterval={pricePerinterval}
+        secondsPerPayment={secondsPerPayment} />
+      :
+      <TimerForm
+        setConfirmed={value => setConfirmed(value)}
+        setPricePerinterval={value => setPricePerinterval(value)}
+        setSecondsPerPayment={value => setSecondsPerPayment(value)} />
 
-  onTimerUpdate = () => {
-    // update timer by 1
-    this.setState({ seconds: this.state.seconds + 1 });
-  }
-  onTimerPlay = () => {
-    // set the timers intervals
-    this.f = setInterval(this.onTimerUpdate, 1000);
-    this.setState({ isTimerRunning: true })
-  }
-  onTimerPause = () => {
-    // clear the timer from ticking
-    clearInterval(this.f);
-    this.setState({ isTimerRunning: false })
-  }
-  onTimerReset = () => {
-    this.onTimerPause()
-    // resets the timer 
-    this.setState({ seconds: 0 })
-  }
-
-  onTimerToggle = () => {
-    if (this.state.isTimerRunning) {
-      this.onTimerPause()
-    } else {
-      this.onTimerPlay()
-    }
-  }
-
-  render() {
-    return (
-      <div className='timer'>
-        <CircularProgressbarWithChildren value={this.getPercentage()} maxValue={1}
-          styles={{
-            path: {
-              stroke: 'rgba(62, 152, 199, 255)',
-              strokeLinecap: 'butt',
-              transition: 'stroke-dashoffset 0.1s ease 0s',
-            },
-            trail: {
-              stroke: '#333',
-              strokeLinecap: 'butt',
-            },
-          }}>
-          <p>{getTimeFormated(this.state.seconds, true)}</p>
-          <p>{`£${this.getAccumulativePrice()} @ £${this.pricePerInterval}/${getTimeFormated(this.secondsPerPayment, false)}`}</p>
-          <div className='buttons'>
-            <button alt='toggle Timer' onClick={this.onTimerToggle}><img className='button-icon' src={this.state.isTimerRunning ? process.env.PUBLIC_URL + '/pause.png' : process.env.PUBLIC_URL + '/play.png'}/></button>
-            <button als='stop timer' onClick={this.onTimerReset} disabled={this.state.seconds===0}><img className='button-icon' src={process.env.PUBLIC_URL + '/stop.png'} /></button>
-          </div>
-
-        </CircularProgressbarWithChildren>
-      </div>
-    )
-  }
+  )
 }
 
 export default Timer
